files upload
This repository contains a project for real-time facial emotion detection using OpenCV and a Convolutional Neural Network (CNN) model trained with Keras. The system can be used in two modes: training the model on a dataset of facial expressions and displaying the detected emotions in real-time using a webcam feed.